delete user confirmation modal

+++ src/routes/admin/+page.svelte
@@ -44,20 +44,30 @@
         })
     }
 
-    async function deleteUser(user: EntireUser) {
-        const response = await fetch(`/api/user/${user.id}`, {
-            method: "DELETE"
-        })
-
-        if (response.status == HttpCodes.Success) {
-            users = users.filter(u => u.id != user.id)
-        } else {
-            const json: ResponseType = await response.json()
-            toastStore.trigger({
-                message: json.message,
-                background: response.status == 200 ? "variant-filled-success" : "variant-filled-error"
+    const deleteUserModal: (user: EntireUser) => ModalSettings = (user) => ({
+        type: "confirm",
+        title: "Delete User Confirmation",
+        body: `Are you sure you want to delete the user <b>${user.username}</b>?`,
+        async response(confirmed: boolean) {
+            if (!confirmed) return
+            const response = await fetch(`/api/user/${user.id}`, {
+                method: "DELETE"
             })
+
+            if (response.status == HttpCodes.Success) {
+                users = users.filter(u => u.id != user.id)
+            } else {
+                const json: ResponseType = await response.json()
+                toastStore.trigger({
+                    message: json.message,
+                    background: response.status == 200 ? "variant-filled-success" : "variant-filled-error"
+                })
+            }
         }
+    })
+
+    async function deleteUser(user: EntireUser) {
+        modalStore.trigger(deleteUserModal(user))
     }
 
 </script>
